adicionando exceptions

+++ src/main/java/br/com/jussara/apirest/produto/controller/ProdutoController.java
@@ -1,14 +1,18 @@
 package br.com.jussara.apirest.produto.controller;
 
+import br.com.jussara.apirest.produto.exception.ResourceNotFoundException;
 import br.com.jussara.apirest.produto.model.ProdutoModel;
 import br.com.jussara.apirest.produto.repository.ProdutoRepository;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static br.com.jussara.apirest.produto.constantes.Constantes.*;
 
 @RestController
 @RequestMapping(value = "/api")
@@ -27,7 +31,10 @@ public class ProdutoController {
 
     @GetMapping("/produto/{id}")
     @ApiOperation(value = "Consultar produto por ID")
-    public ProdutoModel listarProdutoPorID(@PathVariable(value = "id") long id){
+    public ProdutoModel consultarProdutoPorID(@PathVariable(value = "id") long id){
+        if(produtoRepository.findById(id) == null) {
+            throw new ResourceNotFoundException(MESSAGE_PRODUTO_NAO_ENCONTRADO);
+        }
         return produtoRepository.findById(id);
     }
 
@@ -35,24 +42,37 @@ public class ProdutoController {
     @ApiOperation(value = "Criar um produto")
     @ResponseStatus(value = HttpStatus.CREATED)
     public ProdutoModel criarProduto(@RequestBody ProdutoModel produtoModel){
+        if(!produtoRepository.findById(produtoModel.getId()).isPresent()) {
+            throw new ResourceNotFoundException(MESSAGE_PRODUTO_NAO_PODE_SER_VAZIO_OU_NULO);
+        }
         return produtoRepository.save(produtoModel);
     }
 
     @DeleteMapping("/produto")
     @ApiOperation(value = "Deletar um produto")
-    public void deleteProduto(@RequestBody ProdutoModel produtoModel){
+    public void deletarProduto(@RequestBody ProdutoModel produtoModel){
+        if(!produtoRepository.findById(produtoModel.getId()).isPresent()) {
+            throw new ResourceNotFoundException(MESSAGE_PRODUTO_NAO_ENCONTRADO);
+        }
         produtoRepository.delete(produtoModel);
     }
 
     @DeleteMapping("/produto/{id}")
     @ApiOperation(value = "Deletar um produto por ID")
-    public ProdutoModel deletarProdutoPorID(@PathVariable(value = "id") long id){
+    @ApiResponse(code = org.apache.http.HttpStatus.SC_INTERNAL_SERVER_ERROR, message = MESSAGE_ERRO_INTERNO)
+    public ProdutoModel deletarProdutoPorID(@PathVariable(value = "id") Integer id){
+        if(produtoRepository.findById(id) == null) {
+            throw new ResourceNotFoundException(MESSAGE_PRODUTO_NAO_ENCONTRADO);
+        }
         return produtoRepository.deleteById(id);
     }
     
     @PutMapping("/produto")
     @ApiOperation(value = "Atualizar um produto")
     public ProdutoModel atualizarProduto(@RequestBody ProdutoModel produtoModel){
+        if(!produtoRepository.findById(produtoModel.getId()).isPresent()) {
+            throw new ResourceNotFoundException(MESSAGE_PRODUTO_NAO_ENCONTRADO);
+        }
         return produtoRepository.save(produtoModel);
     }
